Move BindingPathUtilities to GitHub.Exports
This is a helper class and not fundamental to GitHub for Visual Studio.
