+++ src/GitHub.VisualStudio/GitHub.VisualStudio.csproj
@@ -330,7 +330,6 @@
       <Link>Properties\SolutionInfo.cs</Link>
     </Compile>
     <Compile Include="AssemblyResolverPackage.cs" />
-    <Compile Include="BindingPathUtilities.cs" />
     <Compile Include="Commands\BlameLinkCommand.cs" />
     <Compile Include="Commands\CreateGistCommand.cs" />
     <Compile Include="Commands\OpenFromClipboardCommand.cs" />
